requirements.txt
pillow>=10.0.0
mss>=9.0.0
anthropic>=0.18.0
imagehash>=4.3.1
pynput>=1.7.6
ddgs>=8.0.0
python-dotenv>=1.0.0
pystray>=0.19.5
pytesseract>=0.3.10
